0x0F-python-object_relational_mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,18 +1,21 @@
 #!/usr/bin/python3
-"""Script that deletes all State objects with a name containing the letter a"""
+# Deletes all State objects with a name containing
+# Usage: ./13-model_state_delete_a.py <mysql username> /
+#                                     <mysql password> /
+#                                     <database name>
 import sys
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
-
+from model_state import State
 
 if __name__ == "__main__":
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'
-                          .format(sys.argv[1], sys.argv[2], sys.argv[3]))
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    states = session.query(State).filter(State.name.like('%a%')).all()
-    for state in states:
-        session.delete(state)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
